Update Includes_Images.xml per jurialmunkey's direction
It's a skin issue due to faulty logic in their watched indicator variables. Specifically this line:
`
<value condition="Integer.IsGreater(ListItem.Property(TotalEpisodes),0) + Integer.IsLess(ListItem.Property(UnWatchedEpisodes),1) + [Container.Content(tvshows) | Container.Content(seasons)]">indicator/checkmark-fo.png</value>
`

`Integer.IsLess` considers `None` to be the same as `0` so this condition marks as watched even though this property doesn't exist for these items.

More importantly, it is faulty logic to consider unwatched episodes being 0 to mean the show is watched. There are many reasons an unwatched show might have 0 unwatched episodes for instance it hasn't aired yet so no episodes exist to watch or there's simply no data available.

A more correct variable condition to use would be the following which checks that the watched episode count matches the total episode count and that there is at least 1 episode.
`
<value condition="Integer.IsGreater(ListItem.Property(TotalEpisodes),0) + String.IsEqual(ListItem.Property(WatchedEpisodes),ListItem.Property(TotalEpisodes)) + [Container.Content(tvshows) | Container.Content(seasons)]">indicator/checkmark-fo.png</value>
`

You will need to get the skin author to fix this.

_Originally posted by @jurialmunkey in https://github.com/jurialmunkey/plugin.video.themoviedb.helper/issues/639#issuecomment-1032023506_